Crate xdg_basedir

xdg-basedir is a utility library to make conforming to the XDG basedir specification easier.

Example

#![cfg(unix)]
extern crate xdg_basedir;

#![cfg(unix)]
use xdg_basedir::*;
use std::path::PathBuf;
...
let data_home: PathBuf = try!(get_data_home());
...

Alternate implementation and some initial source borrowed from rust-xdg. The APIs provided by rust-xdg and xdg-basedir are different.

Reexports

pub use dirs::*;

pub use error::*;

Modules

dirs

error